Add check to not install on linux when conda is already installed

+++ scripts/utils-linux.sh
@@ -23,8 +23,10 @@ function installDependencies() {
   sudo apt update --yes
   sudo apt upgrade --yes
 
-  # Install miniconda
-  wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda.sh
-  bash ~/miniconda.sh -b -p ~/miniconda
-  rm ~/miniconda.sh
+  if [[ ! command -v COMMAND &> /dev/null ]]; then
+    # Install miniconda
+    wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda.sh
+    bash ~/miniconda.sh -b -p ~/miniconda
+    rm ~/miniconda.sh
+  fi
 }
